+++ .travis/deploy_script.sh
@@ -107,9 +107,12 @@ if [ "$TRAVIS" = "true" ]; then
   make clean
 
   cd /tmp
+  ls -al $TRAVIS_BUILD_DIR/../
+
   tar -czf $TAR_NAME -C $TRAVIS_BUILD_DIR/../
   sha256sum $TAR_NAME >"$TAR_NAME.sha256"
   tar2zip $TAR_NAME
+  sha256sum $ZIP_NAME >"$ZIP_NAME.sha256"
 
   ls -al /tmp
 
@@ -117,8 +120,8 @@ if [ "$TRAVIS" = "true" ]; then
   mkdir -p $TRAVIS_BUILD_DIR/build
   cd $TRAVIS_BUILD_DIR/build
 
-  mv "/tmp/$TAR_NAME" "/tmp/$ZIP_NAME" "/tmp/$TAR_NAME.sha256" .
-  artifacts upload --target-paths "/" $ZIP_NAME $TAR_NAME $TRAVIS_BUILD_DIR/VERSION $TAR_NAME.sha256
+  mv "/tmp/$TAR_NAME" "/tmp/$ZIP_NAME" "/tmp/$TAR_NAME.sha256" "/tmp/$ZIP_NAME.sha256".
+  artifacts upload --target-paths "/" $ZIP_NAME $TAR_NAME $TRAVIS_BUILD_DIR/VERSION $TAR_NAME.sha256 $ZIP_NAME.sha256
 
   # Upload the version for a tagged release. Move to a version file in different
   # tags. Within phpVMS, we have an option of which version to track in the admin
