+++ app/Providers/DirectiveServiceProvider.php
@@ -1,0 +1,25 @@
+<?php
+
+namespace App\Providers;
+
+use Illuminate\Support\Facades\Blade;
+use Illuminate\Support\ServiceProvider;
+
+/**
+ * Keep custom directives that can be used in templates
+ *
+ * @package App\Providers
+ */
+class DirectiveServiceProvider extends ServiceProvider
+{
+    public function boot(): void
+    {
+        Blade::directive('minutestotime', function ($expr) {
+            return "<?php echo \App\Support\Units\Time::minutesToTimeString($expr); ?>";
+        });
+
+        Blade::directive('minutestohours', function ($expr) {
+            return "<?php echo \App\Support\Units\Time::minutesToHours($expr); ?>";
+        });
+    }
+}
