+++ resources/views/layouts/default/users/table.blade.php
@@ -41,7 +41,7 @@
         @endif
       </td>
       <td align="center">{{ $user->flights }}</td>
-      <td align="center">{{ \App\Facades\Utils::minutesToTimeString($user->flight_time) }}</td>
+      <td align="center">@minutestotime($user->flight_time)</td>
     </tr>
   @endforeach
   </tbody>
